CA policy in report-only gives a message of not applicable
According to case 120050522001804 and ICM https://portal.microsofticm.com/imp/v3/incidents/details/187386043/home if the user configures the CA policy to report-only mode, the message "Not Applied" is shown to the CX giving the false sensation that if the policy is enabled, the user won't be blocked from external accesses. As explained by the PG, this is by design and in this case, we consider this change in the documentation to have users aware about this CA policy in report-only mode.